upd: マシン側が張力計からjsonで張力を受信するよう改良。実機試験はまだ

+++ tension_meter_machine/BTTensionMeter.hpp
@@ -34,6 +34,6 @@ bool tensionAvailable();
 
 /**
  * @brief Get the latest tension sent from meter.
- * @return (int) Measured tension [mg]
+ * @return (float) Measured tension [kg]
  */
-int getTension();
+float getTension();
